Counter når en melding resendes grunnet manglende ack
