+++ ebms-async/src/main/kotlin/no/nav/emottak/ebms/async/App.kt
@@ -71,32 +71,6 @@ import kotlin.concurrent.timer
 
 val log = LoggerFactory.getLogger("no.nav.emottak.ebms.async.App")
 
-const val MESSAGES_QUEUED_FOR_RETRY_COUNTER = "messages_queued_for_retry_total"
-const val TAG_RETRY_TOPIC = "retry_topic"
-
-const val MESSAGES_FIRST_FAILURE_COUNTER = "messages_first_failure_total"
-const val TAG_DIRECTION = "direction"
-const val TAG_SERVICE = "service"
-const val TAG_ACTION = "action"
-
-fun MeterRegistry.incrementMessagesQueuedForRetry(topic: String) =
-    counter(
-        MESSAGES_QUEUED_FOR_RETRY_COUNTER,
-        TAG_RETRY_TOPIC,
-        topic
-    ).increment()
-
-fun MeterRegistry.incrementFirstFailure(direction: String, service: String, action: String) =
-    counter(
-        MESSAGES_FIRST_FAILURE_COUNTER,
-        TAG_DIRECTION,
-        direction,
-        TAG_SERVICE,
-        service,
-        TAG_ACTION,
-        action
-    ).increment()
-
 fun main() = SuspendApp {
     val config = config()
 
@@ -205,7 +179,8 @@ fun main() = SuspendApp {
             launchMesssageResendTask(
                 config = config,
                 messagePendingAckRepository = messagePendingAckRepository,
-                payloadMessageForwardingService = payloadMessageForwardingService
+                payloadMessageForwardingService = payloadMessageForwardingService,
+                meterRegistry = appMicrometerRegistry
             )
 
             server(
@@ -375,7 +350,8 @@ fun CoroutineScope.launchErrorRetryTaskOutgoing(
 fun CoroutineScope.launchMesssageResendTask(
     config: Config,
     messagePendingAckRepository: MessagePendingAckRepository,
-    payloadMessageForwardingService: PayloadMessageForwardingService
+    payloadMessageForwardingService: PayloadMessageForwardingService,
+    meterRegistry: MeterRegistry
 ) {
     timer(
         name = "Resend Messages Timer",
@@ -390,6 +366,11 @@ fun CoroutineScope.launchMesssageResendTask(
                 log.info("Found ${messagesToResend.size} messages to be resent because of missing Ack")
                 for (message in messagesToResend) {
                     payloadMessageForwardingService.resendMessage(message)
+                    meterRegistry.incrementMessagesResent(
+                        resentCount = message.resentCount + 1,
+                        service = message.messageHeader.service.value ?: "unknown",
+                        action = message.messageHeader.action ?: "unknown"
+                    )
                     messagePendingAckRepository.markResent(message)
                 }
             } catch (e: Exception) {
