+++ ebms-async/src/test/kotlin/no/nav/emottak/ebms/async/LocalApp.kt
@@ -203,7 +203,8 @@ fun main() = SuspendApp {
             launchMesssageResendTask(
                 config = config,
                 messagePendingAckRepository = messagePendingAckRepository,
-                payloadMessageForwardingService = payloadMessageForwardingService
+                payloadMessageForwardingService = payloadMessageForwardingService,
+                meterRegistry = appMicrometerRegistry
             )
 
             server(
